backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
aiosqlite==0.20.0
alembic==1.14.0

# Validation
pydantic==2.10.3
pydantic-settings==2.7.0

# HTTP clients
httpx==0.28.1
requests==2.32.3

# Scraping / parsing
beautifulsoup4==4.12.3
lxml==5.3.0

# Image processing
Pillow==11.1.0

# Utilities
python-dotenv==1.0.1
python-slugify==8.0.4
click==8.1.7
rich==13.9.4
aiofiles==24.1.0
